SGAudit/requirements.txt
django-cors-headers==1.1.0
django==1.10
djangorestframework==3.4.3
django-oauth-toolkit
drfdocs
